Update Random Number Generator
Added a loop to allow the user to continue using the generator over and over until they command the program to "quit". This adds more utility as the user no longer has to reload the web page to use the generator multiple times, but doesn't created an endless loop. Also enhanced formatting so that it is easier to read.